Fix linux filesystem cache calculation
Linux's "Cached" statistic in /proc/meminfo is highly misleading as it
contains swap-backed shmem. This is hard to reclaim at best, and not
at all reclaimable on systems without swap. It should be considered
actual_used, not actual_free.

A more reliable filesystem cache footprint can be calculated from the
file LRUs and the reclaimable slab cache, which is mostly filesystem
metadata like dentries and inodes.